one http status class.
