--- packages/serverpod_client/lib/src/http/http_status_io.dart
@@ -1,1 +1,0 @@
-export 'dart:io' show HttpStatus;
