+++ tests/serverpod_test_client/lib/serverpod_test_client.dart
@@ -1,4 +1,3 @@
 export 'src/protocol/protocol.dart';
 export 'package:serverpod_client/serverpod_client.dart';
-export 'src/http/http_status_html.dart'
-    if (dart.library.io) 'src/http/http_status_io.dart';
+export 'package:serverpod_client/src/http/http_status.dart';
